Refactor the script into a package
