+++ package/setup.cfg
@@ -1,0 +1,24 @@
+[metadata]
+author = Screwdriver.cd team
+author_email = dhubbard@oath.com
+description-file = ../README.md
+name = screwdriver_in_a_box
+url = https://git.ouroath.com/dhubbard/in-a-box
+version = 0.0.0
+
+[options]
+packages =
+    screwdriver_in_a_box
+
+# Add abstract package dependencies here (do not specify specific versions)
+# for more information see https://yahoo.jiveon.com/groups/python/blog/2015/12/15/unconfusing-python-package-dependencies
+install_requires =
+    python_version>="2.7"
+
+[options.extras_require]
+test =
+    vcrpy-unittest
+
+[options.entry_points]
+console_scripts =
+    screwdriver_in_a_box=screwdriver_in_a_box:main
