+++ package/tox.ini
@@ -1,0 +1,15 @@
+[tox]
+envlist = py27,py34,py35,py36
+skip_missing_interpreters=True
+
+[testenv]
+deps =
+    pytest
+    pytest-cov
+
+commands =
+    pytest --cov=screwdriver_in_a_box --cov-report=xml:cobertura.xml --cov-report term-missing tests/
+    # pytest tests
+
+extras =
+    test
